Factorial Finder

+++ src/IterationControlLoopAndJumpStatements/ForLoop/FactorialFinder.java
@@ -12,7 +12,7 @@ public class FactorialFinder {
         System.out.print("Please enter the number: ");
         int num = sc.nextInt();
 
-        int fact = 1;
+        double fact = 1;
 
         for (int i = 1; i <= num; i++) {
             fact = fact * i;
